canvasを含むelementを切り出し
