+++ src/tsx/canvas_with_caption.tsx
@@ -1,0 +1,28 @@
+import { CSSProperties } from "@material-ui/core/styles/withStyles"
+import React from "react"
+import { CaptionCard } from "../tsx/caption_card"
+import { ScreenShotButton } from "../tsx/screen_shot_button"
+
+interface Props {
+  title: string
+  subtitle: JSX.Element
+  body: JSX.Element
+}
+
+export function CanvasWithCaption({ title, subtitle, body }: Props) {
+  const fixedPositionStyle: CSSProperties = { "position": "fixed" }
+  const relativePositionStyle: CSSProperties = { "position": "relative" }
+
+  return (
+    <div>
+      <div style={fixedPositionStyle}>
+        <div id="canvas-parent"></div>
+        <div style={relativePositionStyle}>
+          <ScreenShotButton />
+        </div>
+      </div>
+      <br />
+      <CaptionCard title="BlindPainter" subtitle={subtitle} body={body} />
+    </div>
+  )
+}
